rename SingleAsyncWork to AsyncCondition
also cleans up its API to fix #1933

as part of this refactoring, splits out non-stream-related functionality
from stream.jl and creates libuv.jl and event.jl to contain it
