+++ base/exports.jl
@@ -23,6 +23,7 @@ export
     Serializer,
     Docs,
     Markdown,
+    Threads,
 
 # Types
     AbstractChannel,
