+++ doc/manual/calling-c-and-fortran-code.rst
@@ -1029,14 +1029,18 @@ Some C libraries execute their callbacks from a different thread, and
 since Julia isn't thread-safe you'll need to take some extra
 precautions. In particular, you'll need to set up a two-layered
 system: the C callback should only *schedule* (via Julia's event loop)
-the execution of your "real" callback. To do this, you pass a function
-of one argument (the ``AsyncWork`` object for which the event was
-triggered, which you'll probably just ignore) to ``SingleAsyncWork``::
+the execution of your "real" callback.
+To do this, create a ``AsyncCondition`` object and wait on it::
 
-  cb = Base.SingleAsyncWork(data -> my_real_callback(args))
+  cond = Base.AsyncCondition()
+  wait(cond)
 
 The callback you pass to C should only execute a :func:`ccall` to
-``:uv_async_send``, passing ``cb.handle`` as the argument.
+``:uv_async_send``, passing ``cb.handle`` as the argument,
+taking care to avoid any allocations or other interactions with the Julia runtime.
+
+Note that events may be coalesced, so multiple calls to uv_async_send
+may result in a single wakeup notification to the condition.
 
 More About Callbacks
 --------------------
