+++ doc/stdlib/base.rst
@@ -1233,6 +1233,18 @@ Events
 
    Create a timer that wakes up tasks waiting for it (by calling ``wait`` on the timer object) at a specified interval.  Times are in seconds.  Waiting tasks are woken with an error when the timer is closed (by ``close``\ ). Use ``isopen`` to check whether a timer is still active.
 
+.. function:: AsyncCondition()
+
+   .. Docstring generated from Julia source
+
+   Create a async condition that wakes up tasks waiting for it (by calling ``wait`` on the object) when notified from C by a call to uv_async_send. Waiting tasks are woken with an error when the object is closed (by ``close``\ ). Use ``isopen`` to check whether it is still active.
+
+.. function:: AsyncCondition(callback::Function)
+
+   .. Docstring generated from Julia source
+
+   Create a async condition that calls the given ``callback`` function. The ``callback`` is passed one argument, the async condition object itself.
+
 Reflection
 ----------
